requirements.txt
Langchain
openai
streamlit
pypdf2
python-dotenv
tiktoken
faiss-cpu
streamlit-option-menu==0.3.2
streamlit-extras==0.2.7
streamlit-javascript

